+++ appinfo/routes.php
@@ -82,6 +82,8 @@ return [
 
 		// Datasource
 		['name' => 'datasource#index', 'url' => '/datasource', 'verb' => 'GET'],
+		['name' => 'datasource#own', 'url' => '/datasource/own', 'verb' => 'GET'],
+		['name' => 'datasource#registered', 'url' => '/datasource/registered', 'verb' => 'GET'],
 		['name' => 'datasource#indexFiltered', 'url' => '/datasource/{datasourceType}', 'verb' => 'GET'],
 
 		// Share